pyproject.toml
[project]
name = "oura-hrv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.4.7",
    "polars>=1.33.1",
    "pydantic>=2.11.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
